Move RecommendedUsers component to component folder
